MOSYNC-2331: I forgot these.
